@@ -1,12 +1,10 @@
-import { AfterViewInit, ChangeDetectionStrategy, Component, NgModuleRef, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ChangeDetectionStrategy, Component, NgModuleRef } from '@angular/core';
+import { Router } from '@angular/router';
 import { TaskService, UserService } from '~global-services';
-import { ERM, Task } from '~models';
 import { DialogService } from '~shared/dialog';
-import { FilterType, SearchService } from '~shared/filters';
-import { ListPageComponent } from '~shared/list-page/list-page.component';
+import { SearchService } from '~shared/filters';
 import { SelectionService } from '~shared/list-page/selection.service';
-import { CreateTaskDialogComponent } from '~shared/task/components/create-task-dialog/create-task-dialog.component';
+import { AbstractTaskCommonComponent } from '~shared/task-common/containers/abstract-task-common.component';
 
 @Component({
 	selector: 'product-tasks-app',
@@ -14,31 +12,21 @@ import { CreateTaskDialogComponent } from '~shared/task/components/create-task-d
 	styleUrls: ['./product-tasks.component.scss'],
 	changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ProductTasksComponent extends ListPageComponent<Task, TaskService> implements OnInit, AfterViewInit {
+export class ProductTasksComponent extends AbstractTaskCommonComponent {
 
 	constructor(
-		private userSrv: UserService,
-		private route: ActivatedRoute,
+		protected userSrv: UserService,
 		protected router: Router,
 		protected featureSrv: TaskService,
 		protected searchSrv: SearchService,
 		protected selectionSrv: SelectionService,
 		protected dlgSrv: DialogService,
 		protected moduleRef: NgModuleRef<any>) {
-		super(router, featureSrv, selectionSrv, searchSrv, dlgSrv, moduleRef, ERM.TASK, CreateTaskDialogComponent);
+		super(router, userSrv, featureSrv, searchSrv, selectionSrv, dlgSrv, moduleRef);
 	}
 
-	ngAfterViewInit() {
-		this.filterList.addFilter({ type: FilterType.PRODUCT, value: this.route.parent.snapshot.params.id });
+	createTask(name: string) {
+		// override with the task created and this product
 	}
-
-	toggleFilter(show: boolean) {
-		const filterAssignee = { type: FilterType.ASSIGNEE, value: this.userSrv.userSync.id };
-		if (show) {
-			this.filterList.addFilter(filterAssignee);
-		} else {
-			this.filterList.removeFilter(filterAssignee);
-		}
-	}
-
 }
+
