@@ -31,7 +31,6 @@ import {
 } from '~features/products/components';
 import { NewProductDialogComponent } from '~features/products/components/new-product-dialog/new-product-dialog.component';
 import { ProductDetailsComponent, ProductGeneralInfoComponent, ProductsPageComponent } from '~features/products/containers';
-import { ProductPreviewComponent } from '~features/products/components/product-preview/product-preview.component';
 import { ProductNavigationComponent } from '~features/products/components/product-navigation/product-navigation.component';
 import { WorkflowActionModule } from '~shared/workflow-action/workflow-action.module';
 import { StageIndicatorModule } from '~shared/stage-indicator/stage-indicator.module';
@@ -48,7 +47,7 @@ import { ActivityModule } from '~shared/activity/activity.module';
 import { StatusModule } from '~shared/status/status.module';
 import { VoteDetailsDialogComponent } from './components/vote-details-dialog/vote-details-dialog.component';
 import { ProductTasksComponent } from './containers/product-tasks/product-tasks.component';
-import { TaskModule } from '~shared/task';
+import { TaskCommonModule } from '~shared/task-common';
 
 @NgModule({
 	imports: [
@@ -74,7 +73,7 @@ import { TaskModule } from '~shared/task';
 		StatusModule,
 		ProductCommonModule,
 		UtilsModule,
-		TaskModule
+		TaskCommonModule
 	],
 	declarations: [
 		ProductSmallCardComponent,
@@ -89,7 +88,6 @@ import { TaskModule } from '~shared/task';
 		ProductSummaryComponent,
 		SelectableImageComponent,
 		NewProductDialogComponent,
-		ProductPreviewComponent,
 		ProductNavigationComponent,
 		RfqDialogComponent,
 		EmailListComponent,
