@@ -95,7 +95,11 @@ export class ProjectWorkflowComponent extends AutoUnsub implements OnInit {
 
 	/** Unselects a entity */
 	onAllItemsUnselected(entity: any) {
-		this.selectionSrv.unselectOne(entity);
+		if (Array.isArray(entity)) {
+			entity.forEach(e => this.selectionSrv.unselectOne(e));
+		} else {
+			this.selectionSrv.unselectOne(entity);
+		}
 	}
 
 	/**
