@@ -1,38 +1,27 @@
-import { Component, NgModuleRef, OnInit } from '@angular/core';
+import { Component, NgModuleRef } from '@angular/core';
 import { Router } from '@angular/router';
 import { TaskService, UserService } from '~global-services';
-import { ERM, Task } from '~models';
 import { DialogService } from '~shared/dialog';
-import { FilterType, SearchService } from '~shared/filters';
-import { ListPageComponent } from '~shared/list-page/list-page.component';
+import { SearchService } from '~shared/filters';
 import { SelectionService } from '~shared/list-page/selection.service';
-import { CreateTaskDialogComponent } from '~shared/task/components/create-task-dialog/create-task-dialog.component';
+import { AbstractTaskCommonComponent } from '~shared/task-common/containers/abstract-task-common.component';
 
 @Component({
 	selector: 'workspace-my-tasks-page-app',
 	templateUrl: './my-tasks-page.component.html',
 	styleUrls: ['./my-tasks-page.component.scss']
 })
-// the service should be TaskService instead ofthis temporary one
-export class MyTasksPageComponent extends ListPageComponent<Task, TaskService> implements OnInit {
+export class MyTasksPageComponent extends AbstractTaskCommonComponent {
 
 	constructor(
-		private userSrv: UserService,
+		protected userSrv: UserService,
 		protected router: Router,
 		protected featureSrv: TaskService,
 		protected searchSrv: SearchService,
 		protected selectionSrv: SelectionService,
 		protected dlgSrv: DialogService,
 		protected moduleRef: NgModuleRef<any>) {
-		super(router, featureSrv, selectionSrv, searchSrv, dlgSrv, moduleRef, ERM.TASK, CreateTaskDialogComponent);
+		super(router, userSrv, featureSrv, searchSrv, selectionSrv, dlgSrv, moduleRef);
 	}
 
-	toggleFilter(show: boolean) {
-		const filterAssignee = { type: FilterType.ASSIGNEE, value: this.userSrv.userSync.id };
-		if (show) {
-			this.filterList.addFilter(filterAssignee);
-		} else {
-			this.filterList.removeFilter(filterAssignee);
-		}
-	}
 }
