@@ -126,14 +126,14 @@ export class WorkspaceFeatureService extends ProductService {
 	updateProductStatus(product: Product, statusType: ProductStatusType) {
 		// we check if the product has a status
 		if (!product.status) {
-			const tempStatus = new ProductStatus({ status: { id: statusType.id } }) as any;
-			return this.update({ id: product.id, status: tempStatus }, [ProductQueries.status]);
+			const tempStatus = new ProductStatus({ status: statusType }) as any;
+			return this.update({ id: product.id, status: tempStatus });
 		} else {
 			// we dont update if we click the same status as the current one of the product
 			const productStatusType = product.status.status;
 			if (statusType.id !== productStatusType.id) {
-				const tempStatus = new ProductStatus({ status: { id: statusType.id } }) as any;
-				return this.update({ id: product.id, status: tempStatus }, [ProductQueries.status]);
+				const tempStatus = new ProductStatus({ status: statusType }) as any;
+				return this.update({ id: product.id, status: tempStatus });
 			}
 		}
 		return of();
