@@ -1,13 +1,12 @@
-import { DocumentNode } from "graphql";
-import { Observable } from "rxjs";
-import { QueryRef } from "apollo-angular";
-import { SelectParamsConfig } from "~global-services/_global/select-params";
+import { DocumentNode } from 'graphql';
+import { Observable } from 'rxjs';
+import { QueryRef } from 'apollo-angular';
+import { SelectParamsConfig } from '~global-services/_global/select-params';
 
 
 export interface ListQuery<T> {
 	queryName: string;
-	queryRef: QueryRef<any, any>;
 	items$: Observable<T[]>;
-	fetchMore: (skip: number) => void;
-	refetch: (params: SelectParamsConfig) => void;
+	fetchMore: (skip: number) => Observable<any>;
+	refetch: (params: SelectParamsConfig) => Observable<any>;
 }
