@@ -25,7 +25,6 @@ export class UserClientInitializer extends AbstractApolloClient {
 		protected apolloState: ApolloStateService,
 		protected userSrv: UserService,
 		protected realmServerSrv: RealmServerService
-
 	) {
 		super(apollo, link, apolloState, realmServerSrv);
 	}
@@ -35,6 +34,7 @@ export class UserClientInitializer extends AbstractApolloClient {
 		// we get the user id from the auth service
 		const userId$ = this.authSrv.userId$.pipe(
 			filter(id => !!id),
+			tap(_ => this.apolloState.setClientPending(Client.USER)),
 			// we don't use distinctUntilChanged because an user can reconnect with the same account
 			shareReplay(1)
 		);
@@ -48,35 +48,26 @@ export class UserClientInitializer extends AbstractApolloClient {
 
 		const realmUri$ = userId$.pipe(
 			// realm uri won't change if the userId hasn't changed
-			distinctUntilChanged(),
-			switchMap(userId => this.getUserRealmUri(userId)),
+			switchMap(userId => this.getUserRealmUri(userId))
 		);
 
-		combineLatest(realmUri$, accessToken$)
+		zip(realmUri$, accessToken$)
 			.subscribe(([uri, token]) => super.initClient(uri, Client.USER, token));
 
 
 		// when the refreshToken is gone we close it
 		this.authSrv.authStatus$.pipe(
 			filter(status => status === AuthStatus.NOT_AUTHENTICATED),
-		).subscribe(_ => this.destroyClient(Client.USER));
-
+		).subscribe(_ => this.destroyClient(Client.USER, 'not authenticated'));
 	}
 
 	/** will emit once when all user and global constant are ready */
-	private requiredClientsReady() {
-		// we need to wait for all user client and global const client to be ready
-		return this.apolloState
-			.getClientStatus(Client.ALL_USER).pipe(
-				filter(status => status === ClientStatus.READY),
-		);
-	}
-
 	private getUserRealmUri(userId: string) {
 		// then we can query the user, and with that user we can get the realm uri...
-		return this.requiredClientsReady().pipe(
-			switchMap(_ => this.userSrv.queryOne(userId, 'realmServerName, realmPath', Client.ALL_USER).pipe(first())),
-			switchMap(user => super.getRealmUri(user.realmServerName, user.realmPath)),
+		return this.userSrv.queryOne(userId, 'realmServerName, realmPath', Client.ALL_USER)
+			.pipe(
+				first(),
+				switchMap(user => super.getRealmUri(user.realmServerName, user.realmPath)),
 		);
 	}
 
