@@ -83,9 +83,9 @@ export class ProductPreviewComponent extends AutoUnsub implements OnInit {
 			takeUntil(this._destroy$),
 			map(product => new FormDescriptor(this.customFields2, product))
 		).subscribe(this.descriptor2$);
-		this.featureSrv.getContacts(this.product.supplier.id).pipe(
+		/* this.featureSrv.getContacts(this.product.supplier.id).pipe(
 			takeUntil(this._destroy$)
-		).subscribe(supp => this.contacts = supp.contacts);
+		).subscribe(supp => this.contacts = supp.contacts); */
 	}
 
 	/** when we receive back the form from the dynamic form component we subscribe to changes to it and
@@ -104,13 +104,10 @@ export class ProductPreviewComponent extends AutoUnsub implements OnInit {
 	}
 
 	openRfq() {
+		// TODO: contacts must be loaded from the RfqDialog itself
 		// we add manually the supplier self email, since it is not on the contacts
-		if (this.contacts && this.product.supplier.officeEmail) {
-			this.contacts.push({
-				name: this.product.supplier.name || 'Unnamed',
-				email: this.product.supplier.officeEmail,
-				jobTitle: null
-			});
+		/* if (this.contacts && this.product.supplier.officeEmail) {
+			this.contacts.push({ name: this.product.supplier.name || 'Unnamed', email: this.product.supplier.officeEmail, jobTitle: null });
 		} else if (!this.contacts && this.product.supplier.officeEmail) {
 			this.contacts = [{
 				name: this.product.supplier.name || 'Unnamed',
@@ -123,6 +120,7 @@ export class ProductPreviewComponent extends AutoUnsub implements OnInit {
 				product: this.product,
 				contacts: this.contacts
 			});
+		this.dlgSrv.openFromModule(RfqDialogComponent, this.module, { product: this.product, contacts: this.contacts }); */
 	}
 
 	onViewProduct() {
