@@ -23,14 +23,12 @@ export class ThumbButtonsComponent extends AutoUnsub implements OnInit {
 		this.userVoteIndex = (votes || []).findIndex(v => v.user.id === this.userSrv.userSync.id);
 		if (~this.userVoteIndex)
 			this.userVote = votes[this.userVoteIndex];
-		this._votes = votes;
+		this._votes = votes || [];
 	}
 	get votes() {
 		return this._votes;
 	}
 	userVoteIndex: number;
-
-
 
 	@Input() size = 's';
 	// when we want the color of the thumb be the background isntead of the icon
@@ -131,7 +129,7 @@ export class ThumbButtonsComponent extends AutoUnsub implements OnInit {
 	}
 
 	updateEmitVote() {
-		const value = this.userVote.value === 100 ? 0 : 100;
+		const value = (!this.userVote || this.userVote.value === 100) ? 0 : 100;
 		this.userVote = { ...this.userVote, value };
 		this._votes = [...this._votes];
 		this._votes[this.userVoteIndex] = this.userVote;
