@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { UserPictureModule } from '~shared/user-picture';
 import { UtilsModule } from '~shared/utils';
-import { TaskListComponent } from '~shared/task/components/task-list/task-list.component';
-import { TaskComponent } from '~shared/task/components/task/task.component';
+import { TaskListComponent } from '~shared/task-common/components/task-list/task-list.component';
+import { TaskComponent } from '~shared/task-common/components/task/task.component';
 import { ImageModule } from '~shared/image/image.module';
 import { InputsModule } from '~shared/inputs';
 import { IconsModule } from '~shared/icons';
@@ -30,4 +30,4 @@ import { DynamicFormsModule } from '~shared/dynamic-forms';
 	exports: [TaskListComponent, TaskComponent, CreateTaskDialogComponent, PreviewTaskComponent],
 	entryComponents: [CreateTaskDialogComponent]
 })
-export class TaskModule { }
+export class TaskCommonModule { }
